Rename `custom_logger` to `logging`
