+++ search_engines/text/base.py
@@ -1,6 +1,6 @@
 from typing import Iterator
 
-from utils.custom_logger import main_logger
+from utils.logging import main_logger
 
 
 class TextSearchEngine:
