+++ utils/proxy_getter.py
@@ -2,7 +2,7 @@ import threading
 
 import proxyscrape
 
-from utils.custom_logger import main_logger
+from utils.logging import main_logger
 
 
 class ProxyGetter:
